requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1

# AI and LLM
google-generativeai
langchain
langchain-google-genai

# File processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
python-docx==1.1.0

# NLP and text processing
spacy==3.8.7
sentence-transformers==5.1.0
scikit-learn==1.3.2
nltk==3.9.1

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.1
pydantic==2.11.9
pydantic-settings==2.6.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1